feat(database): PostgreSQL dialect & SQL compatibility
Make the existing query/schema layer run on PostgreSQL as well as SQLite:
- fix dialect-sensitive schema/migration types (001/010/016/027) and add a
  Postgres heal migration (032); renumber upstream's nzb_store_refs to 033
- build health-listing WHERE clauses dynamically to avoid untyped
  `? IS NULL` params (Postgres error 42P18)
- fully qualify ON CONFLICT / EXCLUDED column refs across health, queue and
  stats upserts to remove ambiguous-column errors (42702)
- force UTC session timezone on all Postgres connections
- add a dual-engine (SQLite + Postgres) CI matrix and dialect parity tests

Squashed from: 1e3b93e f8be0f3 2ac5a9f 2bc515b cf0fbe1 135a4c9

+++ internal/database/migrations/postgres/001_initial_schema.sql
@@ -105,7 +105,7 @@ CREATE TABLE media_files (
     instance_type TEXT NOT NULL CHECK(instance_type IN ('radarr', 'sonarr')),
     external_id INTEGER NOT NULL,
     file_path TEXT NOT NULL,
-    file_size INTEGER,
+    file_size BIGINT,
     created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
